# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go-Rocks is a tech playground with Go, Gin",
        "title": "Swagger Go-Rocks",
        "contact": {
            "name": "Lingga Edo",
            "url": "http://www.github.com/linggaaskaedo",
            "email": "lemp.otis@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/user": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for creating a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload create user",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Endpoint for generating access token for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Payload user login",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserLoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint user for logging out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload user logout",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserLogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserLogoutResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/user/relog": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint user for renewing access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User relog",
                "parameters": [
                    {
                        "description": "Payload user logout",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UserRelogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserLoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        },
        "/user/{userID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoint for get a user with ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPUserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPErrResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.Meta": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "x-order": "0"
                },
                "status_code": {
                    "type": "integer",
                    "x-order": "1"
                },
                "status": {
                    "type": "string",
                    "x-order": "2"
                },
                "message": {
                    "type": "string",
                    "x-order": "3"
                },
                "error": {
                    "type": "object",
                    "x-order": "4"
                },
                "timestamp": {
                    "type": "string",
                    "x-order": "5"
                }
            }
        },
        "dto.UserDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0"
                },
                "username": {
                    "type": "string",
                    "x-order": "1"
                },
                "email": {
                    "type": "string",
                    "x-order": "2"
                },
                "phone": {
                    "type": "string",
                    "x-order": "3"
                },
                "isDeleted": {
                    "type": "boolean",
                    "x-order": "5"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginDTO": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "x-order": "0"
                },
                "refreshToken": {
                    "type": "string",
                    "x-order": "1"
                },
                "expiresAt": {
                    "type": "integer",
                    "x-order": "2"
                },
                "expiresRt": {
                    "type": "integer",
                    "x-order": "3"
                },
                "expiresIn": {
                    "type": "integer",
                    "x-order": "4"
                }
            }
        },
        "dto.UserLogoutDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "0"
                }
            }
        },
...
        "rest.CreateUserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/rest.UserDataPayload"
                }
            }
        },
...
        "rest.HTTPErrResp": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPUserLoginResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.UserLoginData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPUserLogoutResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.UserLogoutData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.HTTPUserResp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.UserData"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.Meta"
                }
            }
        },
        "rest.LoginUserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/rest.UserLoginDataPayload"
                }
            }
        },
        "rest.LogoutUserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/rest.UserLogoutDataPayload"
                }
            }
        },
        "rest.RelogUserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/rest.UserRelogDataPayload"
                }
            }
        },
        "rest.UserCreateRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.CreateUserData"
                }
            }
        },
        "rest.UserData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/dto.UserDTO"
                }
            }
        },
        "rest.UserDataPayload": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "x-order": "0"
                },
                "email": {
                    "type": "string",
                    "x-order": "1"
                },
                "phone": {
                    "type": "string",
                    "x-order": "2"
                },
                "divisionID": {
                    "type": "string",
                    "x-order": "3"
                },
                "password": {
                    "type": "string",
                    "x-order": "4"
                }
            }
        },
        "rest.UserLoginData": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/dto.UserLoginDTO"
                }
            }
        },
        "rest.UserLoginDataPayload": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "x-order": "0",
                    "example": "ahmad"
                },
                "password": {
                    "type": "string",
                    "x-order": "1",
                    "example": "sahroni"
                }
            }
        },
        "rest.UserLoginRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.LoginUserData"
                }
            }
        },
        "rest.UserLogoutData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserLogoutDTO"
                }
            }
        },
        "rest.UserLogoutDataPayload": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "rest.UserLogoutRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.LogoutUserData"
                }
            }
        },
        "rest.UserRelogDataPayload": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "rest.UserRelogRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.RelogUserData"
                }
            }
        }
    }
}